Enhanced maintainability
